data-analysis-pipeline/requirements.txt
streamlit==1.28.2
pandas==2.1.3
plotly==5.17.0
mysql-connector-python==8.2.0
numpy==1.24.4
python-dotenv==1.0.0
